// src/types/Operaciones/Senia.ts
export interface Senia {
  ID: number;
  CuotaACobrar: number;
  cuentaContable?: string;
  codigoMarca: number;
  seniasFiltered: number;
  Fecha: string;
  Importe: number;
  Interes: number;
  ImpAbonado: number;
  FormaDePago: string;
  Tarjeta: number;
  NroTarjeta: number;
  NroCupon: number;
  FechaCupon: string;
  Lote: number;
  CantPagos: number;
  NroRecibo: string;
  FechaVto: string;
}