// src/types/Operaciones/Interes.ts
export interface Interes {
  Id: number;
  MedioCobro: number;
  Cantidad: number;
  Interes: number;
  Activo: number;
  Codigo: number;
  Nombre: string;
  SolicitaCtaBancaria: number;
  SolicitaNroCheque: number;
  CuentaContable: number;
  UsarEnMesaPlanesCompra: number;
  UsarEnMesaPlanesVta: number;
  UsarEnCompraProv: number;
  UsarEnMesaPlanesCostos: number;
  CtaSecundaria: number;
  C2: number;
  CtaDefCompraMP_C2: number;
  UsarEnHaberesNetosPagoVoucher: number;
  EsTarjeta: number;
  EsEfvo: number;
  UsarEnPreSolicitudes: number;
  SolicitaPreSolicitud: number;
  UsarEnCuotasRecuperoCobro: number;
  UsarEnCuotasRecuperoPago: number;
  EsGanancia: number;
  UsarEnMoraTecnicaCobro: number;
  UsarEnCobranzaFacturasGiama: number;
  BajaACuotaTerminal: number;
}